+++ packages/calendar/src/components/input-types/date-time-input/DateTimeInput.tsx
@@ -1,0 +1,8 @@
+import { TextInput, TextInputProps } from "@stenajs-webui/forms";
+import * as React from "react";
+
+export interface DateTimeInputProps extends TextInputProps {}
+
+export const DateTimeInput: React.FC<DateTimeInputProps> = (props) => (
+  <TextInput type={"datetime-local"} {...props} />
+);