+++ packages/calendar/src/components/input-types/time-input/TimeInput.tsx
@@ -1,0 +1,9 @@
+import { faClock } from "@fortawesome/free-solid-svg-icons/faClock";
+import { TextInput, TextInputProps } from "@stenajs-webui/forms";
+import * as React from "react";
+
+export interface TimeInputProps extends TextInputProps {}
+
+export const TimeInput: React.FC<TimeInputProps> = (props) => (
+  <TextInput iconLeft={faClock} type={"time"} {...props} />
+);